requirements.txt
Django==1.5.5
Markdown==2.3.1
djangorestframework==2.3.9
lxml==3.2.4
wsgiref==0.1.2
